fix(valence): run migration column-naming lint on a live pool before closePool -- GREEN
